Created folder structure for appRelease sharedResource.
