+++ src/react-apps/applications/service-development/src/sharedResources/appRelease/appReleaseDispatcher.ts
@@ -1,29 +1,30 @@
 import { Action, ActionCreatorsMapObject, bindActionCreators } from 'redux';
 import { store } from '../../store';
-import * as AppReleaseActions from './appReleaseActions';
+import * as CreateAppReleaseActions from './create/createAppReleaseActions';
+import * as GetAppReleaseActions from './get/getAppReleasesActions';
 import { IRelease } from './types';
 
 export interface IAppReleaseActionDispatcher extends ActionCreatorsMapObject {
-  getReleases: () => Action;
-  getReleasesFulfilled: (releases: IRelease[]) => AppReleaseActions.IGetReleaseActionFulfilled;
-  getReleasesRejected: (error: Error) => AppReleaseActions.IGetReleaseActionRejected;
-  createRelease: (
+  getAppReleases: () => Action;
+  getAppReleasesFulfilled: (releases: IRelease[]) => GetAppReleaseActions.IGetReleaseActionFulfilled;
+  getAppReleasesRejected: (error: Error) => GetAppReleaseActions.IGetReleaseActionRejected;
+  createAppRelease: (
     tagName: string,
     name: string,
     body: string,
     targetCommitish: string,
-  ) => AppReleaseActions.ICreateReleaseAction;
-  createReleaseFulfilled: (id: string) => AppReleaseActions.ICreateReleaseFulfilledAction;
-  createReleaseRejected: (error: Error) => AppReleaseActions.ICreateReleaseRejectedActions;
+  ) => CreateAppReleaseActions.ICreateReleaseAction;
+  createAppReleaseFulfilled: (id: string) => CreateAppReleaseActions.ICreateReleaseFulfilledAction;
+  createAppReleaseRejected: (error: Error) => CreateAppReleaseActions.ICreateReleaseRejectedActions;
 }
 
 const actions: IAppReleaseActionDispatcher = {
-  getReleases: AppReleaseActions.getReleases,
-  getReleasesFulfilled: AppReleaseActions.getReleasesFulfilled,
-  getReleasesRejected: AppReleaseActions.getReleasesRejected,
-  createRelease: AppReleaseActions.createRelease,
-  createReleaseFulfilled: AppReleaseActions.createReleaseFulfilled,
-  createReleaseRejected: AppReleaseActions.createReleaseRejected,
+  getAppReleases: GetAppReleaseActions.getAppReleases,
+  getAppReleasesFulfilled: GetAppReleaseActions.getAppReleasesFulfilled,
+  getAppReleasesRejected: GetAppReleaseActions.getAppReleasesRejected,
+  createAppRelease: CreateAppReleaseActions.createAppRelease,
+  createAppReleaseFulfilled: CreateAppReleaseActions.createAppReleaseFulfilled,
+  createAppReleaseRejected: CreateAppReleaseActions.createAppReleaseRejected,
 };
 
 const AppReleaseActionDispatcher: IAppReleaseActionDispatcher = bindActionCreators<
